refactor: align view namespaces with root
