+++ ControlesAccesoQR/ViewModels/ControlesAccesoQR/MainWindowViewModel.cs
@@ -11,7 +11,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Threading;
 using ControlesAccesoQR.Models;
-using ControlesAccesoQR.Views.ControlesAccesoQR;
+using ControlesAccesoQR.Views.EntradaSalida;
 using ControlesAccesoQR.Servicios;
 
 using EstadoPanel = ControlesAccesoQR.Estados.EstadoProceso;
