+++ ControlesAccesoQR/Views/EntradaSalida/DialogoHuella.xaml.cs
@@ -1,6 +1,6 @@
 using System.Windows;
 
-namespace ControlesAccesoQR.Views.ControlesAccesoQR
+namespace ControlesAccesoQR.Views.EntradaSalida
 {
     public partial class DialogoHuella : Window
     {
