+++ ControlesAccesoQR/Views/EntradaSalida/DialogoQr.xaml
@@ -1,4 +1,4 @@
-<Window x:Class="ControlesAccesoQR.Views.ControlesAccesoQR.DialogoQr"
+<Window x:Class="ControlesAccesoQR.Views.EntradaSalida.DialogoQr"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Title="Escanear QR" Height="150" Width="300" WindowStartupLocation="CenterOwner"
