+++ ControlesAccesoQR/Views/EntradaSalida/EstadoProcesoPanel.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="ControlesAccesoQR.Views.ControlesAccesoQR.EstadoProcesoPanel"
+<UserControl x:Class="ControlesAccesoQR.Views.EntradaSalida.EstadoProcesoPanel"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:models="clr-namespace:ControlesAccesoQR.Models"
