+++ ControlesAccesoQR/Views/EntradaSalida/VistaSalidaFinal.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="ControlesAccesoQR.Views.ControlesAccesoQR.VistaSalidaFinal"
+<UserControl x:Class="ControlesAccesoQR.Views.EntradaSalida.VistaSalidaFinal"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:uc="clr-namespace:ControlesAccesoQR.UserControls"
